+++ src/main/java/act/handler/builtin/controller/AfterInterceptorInvoker.java
@@ -23,10 +23,11 @@ package act.handler.builtin.controller;
 import act.Destroyable;
 import act.app.ActionContext;
 import act.security.CORS;
+import act.util.Ordered;
 import act.util.Prioritised;
 import org.osgl.mvc.result.Result;
 
-public interface AfterInterceptorInvoker extends Prioritised, Destroyable {
+public interface AfterInterceptorInvoker extends Prioritised, Destroyable, Ordered {
     Result handle(Result result, ActionContext actionContext) throws Exception;
     void accept(ActionHandlerInvoker.Visitor visitor);