+++ src/main/java/act/handler/builtin/controller/RequestHandlerProxy.java
@@ -33,6 +33,7 @@ import act.app.event.SysEventId;
 import act.controller.*;
 import act.controller.meta.*;
 import act.handler.RequestHandlerBase;
+import act.inject.util.Sorter;
 import act.security.CORS;
 import act.security.CSRF;
 import act.util.*;
@@ -203,6 +204,7 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
                 context.enableCache();
             }
             saveActionPath(context);
+            setHandlerClass(context);
             context.startIntercepting();
             result = handleBefore(context);
             if (null == result) {
@@ -349,6 +351,11 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
         context.actionPath(actionPath);
     }
 
+    // in case interceptor class is super type of action handler class
+    private void setHandlerClass(ActionContext context) {
+        context.handlerClass(actionHandler.invoker().invokeMethod().getDeclaringClass());
+    }
+
     private boolean matches(Set<String> patterns) {
         if (patterns.contains(actionMethodName) || patterns.contains(actionPath)) {
             return true;
@@ -441,7 +448,7 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
                 continue;
             }
             BeforeInterceptor interceptor = mode.createBeforeInterceptor(info, app);
-            beforeInterceptors.add(interceptor);
+            insertInterceptor(beforeInterceptors, interceptor);
             sessionFree = sessionFree && interceptor.sessionFree();
             express = express && interceptor.express();
         }
@@ -450,7 +457,7 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
                 continue;
             }
             AfterInterceptor interceptor = mode.createAfterInterceptor(info, app);
-            afterInterceptors.add(interceptor);
+            insertInterceptor(afterInterceptors, interceptor);
             sessionFree = sessionFree && interceptor.sessionFree();
             express = express && interceptor.express();
         }
@@ -459,7 +466,7 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
                 continue;
             }
             ExceptionInterceptor interceptor = mode.createExceptionInterceptor(info, app);
-            exceptionInterceptors.add(interceptor);
+            insertInterceptor(exceptionInterceptors, interceptor);
             sessionFree = sessionFree && interceptor.sessionFree();
             express = express && interceptor.express();
         }
@@ -470,7 +477,7 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
                 continue;
             }
             FinallyInterceptor interceptor = mode.createFinallyInterceptor(info, app);
-            finallyInterceptors.add(interceptor);
+            insertInterceptor(finallyInterceptors, interceptor);
             sessionFree = sessionFree && interceptor.sessionFree();
             express = express && interceptor.express();
         }
@@ -600,7 +607,6 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
 
     public static void registerGlobalInterceptor(ExceptionInterceptor interceptor) {
         insertInterceptor(globalExceptionInterceptors, interceptor);
-        Collections.sort(globalExceptionInterceptors);
     }
 
     @AnnotatedClassFinder(value = Global.class, callOn = SysEventId.PRE_START)
@@ -629,28 +635,11 @@ public final class RequestHandlerProxy extends RequestHandlerBase {
     }
 
     public static <T extends Handler> void insertInterceptor(List<T> list, T i) {
-        int sz = list.size();
-        if (0 == sz) {
-            list.add(i);
-        }
-        ListIterator<T> itr = list.listIterator();
-        while (itr.hasNext()) {
-            T t = itr.next();
-            int n = i.compareTo(t);
-            if (n < 0) {
-                itr.add(i);
-                return;
-            } else if (n == 0) {
-                if (i.equals(t)) {
-                    // already exists
-                    return;
-                } else {
-                    itr.add(i);
-                    return;
-                }
-            }
+        if (list.contains(i)) {
+            return;
         }
         list.add(i);
+        Collections.sort(list, Sorter.COMPARATOR);
     }
 
     public static class GroupInterceptorWithResult {